server/requirement.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
SQLAlchemy==2.0.29
typing_extensions==4.10.0
Werkzeug==3.0.2
redis
pandas